rename report model to cache model
