+++ internal/controller/informer/execution_informer.go
@@ -1,12 +1,12 @@
 package informer
 
 import (
-	"soarca/models/report"
+	"soarca/models/cache"
 
 	"github.com/google/uuid"
 )
 
 type IExecutionInformer interface {
 	GetExecutionsIds() []string
-	GetExecutionReport(executionKey uuid.UUID) (report.ExecutionEntry, error)
+	GetExecutionReport(executionKey uuid.UUID) (cache.ExecutionEntry, error)
 }
