+++ models/cache/cache.go
@@ -1,4 +1,4 @@
-package report
+package cache
 
 import (
 	"soarca/models/cacao"
